- Adds in the beginning of functionality for React login components; - TODO: fill out Login state updating and button event-handling;
