+++ client/src/Components/Login/index.js
@@ -1,12 +1,16 @@
 import React from 'react';
-
+import PasswordForm from '../PasswordForm';
+import UsernameForm from '../UsernameForm';
 
 
 const Login= ()=> {    
     return(
         <div>
             <main>
-                <h1>Login page Coming Soon!</h1>    
+                <h1>You'll soon be able to log in right below this!</h1>
+                  <UsernameForm />
+              {/* eslint-disable-next-line react/jsx-no-undef */}
+                  <PasswordForm />
             </main>
         </div>
         )
